Enable Mac CVDisplayLinkBeginFrameSource by default
With CL 5038373: Disable the feature kDelayOnFramePresent on Mac |
https://chromium-review.googlesource.com/c/chromium/src/+/5038373,
the PresentationFeedback timestamps will be restored to the original
one which is the rendering completed time done instead of the
display time. Now CompositorLatency.TotalLatency and
WebVitals.InteractionToNextPaint2 should not be regressed.

The CVDisplayLinkBeginFrameSource feature will only improves the
accuracy of the BeginFrame callback time for now. It will not change
the Frame Presentation time.

We will try to enable kDelayOnFramePresent and communicate with the
metrics team later on the PresentationFeedback timestamp change.

Bug: 1404797
Change-Id: I3f533f0f66c5fb2c1185d66c5a0ffc587e67a9e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5050323
Reviewed-by: ccameron chromium <ccameron@chromium.org>
Commit-Queue: Maggie Chen <magchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1227649}