Fixing a crash when side panel pinning feature is enabled
The issue seems to be current_entry_ member is reset and an extension
observer call is calling UpdateHeaderPinButtonState. Adding a check for
the member not being nullptr in the function, which is similar to a few
other places.

Bug: 1503672
Change-Id: Ia0fc6f3d7b118d2cbf0c3eda697d0a33aa80944a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5044562
Reviewed-by: Caroline Rising <corising@chromium.org>
Commit-Queue: Eshwar Stalin <estalin@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1227662}