Convert GeneratedHttpsFirstModePref into a tri-state enum
GeneratedHttpsFirstModePref is used for managing the HTTPS-First Mode
setting in the Security settings page. This updates it to match the
HttpsFirstModeSettings enum that adds new states besides "disabled"
and "fully enabled", and hooks it up to the two HTTPS-First Mode prefs
kHttpsOnlyModeEnabled and kHttpsFIrstModeIncognito.

This is an initial step towards adding a new Settings UI (when the
HttpsFirstModeIcnognito feature flag is enabled) to control the
tri-state of HFM (EnabledFull, EnabledIncognito, Disabled). Having
the SettingsToggleButton control the underlying numeric setting will
make it easier to have both versions of the Settings control use the
same underlying generated pref logic.

Bug: 1494186
Change-Id: Iae8dc496f786a60fb5d89bcd6f973eafbd5f7739
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5041962
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Commit-Queue: Chris Thompson <cthomp@chromium.org>
Reviewed-by: Rainhard Findling <rainhard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1227664}