[PWAs] Introduce utility to notify a view becoming visible in an active widget
This CL implements an utility that takes in a widget, an element
identifier and a callback as an input, and posts the callback
to the task runner if the widget is activated and the element is visible. If not, the utility subscribes to changes in the
activation of the widget and the visibility of the element, and
posts the callback once that is met.

This is a fire-and-forget utility that will be tied to the
lifetime of the widget, and is responsible for destroying
itself. This is a part of web apps code as per discussion with views/ OWNERs, since it's only required by web apps code at the moment.

DD: go/views-visibility-activation-waiter
Bug: 1412856, b/305936304
Change-Id: I4a8fce7432089a43afdc5b6eb59280ad96cad2d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5008288
Reviewed-by: Dana Fried <dfried@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Mike Wasserman <msw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1227710}