Split TestAcceptDialogCallback into two to better match prod behavior
TestAcceptDialogCallback was used in FetchManifestAndInstall in several
tests and it hides the fact that the dialog is normally responsible for
setting the UserDisplayMode which in turn affects whether the app is
opened in a window, using display modes from the manifest. Despite tests
using OMNIBOX_INSTALL_ICON, they were installing with UDM::kBrowser.

Ideally we make test behavior match prod behavior better so it's tested.
This CL takes a step toward that by explicitly setting the UDM during
the dialog callback in tests. It also fixes some tests that were broken
by the app popping into a separate window when installed.

Bug: 1503010
Change-Id: I82b34df5d6ef2070660d073847427f12f01800c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5040082
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Commit-Queue: Glen Robertson <glenrob@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1227735}