[QuickStart] Extend response timeout.
A recent change introduced a prompt on the source device for the user
to confirm which account they would like to use before sending a
response to the BootstrapConfigurations message. This causes us to trip
the 3s response timeout.

This extends the default timeout to 1 minute to give us more buffer, and
uses a 30 minute timeout for the BootstrapConfigurations message.

Test: Manually verified on Atlas + Pixel 5a.

Bug: None.
Change-Id: I3f38e85154855df86f7b037fd85f2a503f93901a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5051622
Reviewed-by: Curt Clemens <cclem@google.com>
Commit-Queue: Michael Hansen <hansenmichael@google.com>
Cr-Commit-Position: refs/heads/main@{#1227741}