Add new HTTPS-First Mode tri-state settings UI
This adds a new radio-button selection UI for the different HTTPS-First
Mode settings (fully enabled, enabled in Incognito, and fully disabled)
to chrome://settings/security, when the HttpsFirstModeIncognito feature
is enabled. This changes the previous single SettingsToggleButton row in
the "Advanced" section with a new row (with updated setting title and
subtitle) and a SettingsRadioGroup with ControlledRadioButton elements
for each of the setting states.

This also updates the tests in security_page_test.ts to test both the
new UI (in the "Main" test) and the previous single-toggle UI (in the
"FlagsDisabled" test).

Followup CLs will connect the new UI to the "HttpsOnlyMode" enterprise
policy so that the enterprise policy indicator correctly displays, and
so the "disallowed" policy setting correctly disables HTTPS-First Mode
entirely (including in Incognito).

Screenshot (before):
https://drive.google.com/file/d/1IQxo-_pssD23L0X66l-x7HWsmNJagNqv/view?usp=sharing

Screenshot (after):
https://drive.google.com/file/d/16VTcMgxWrw66Re13O4p8uWwUFHRLVc-7/view?usp=sharing

Bug: 1494186
Change-Id: I108c611560c6e2f46cf3898ac28033b6b18fe711
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5046125
Reviewed-by: Rainhard Findling <rainhard@chromium.org>
Commit-Queue: Chris Thompson <cthomp@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1227757}