Mention the git longpaths support for Windows
By default git is constrained to the 260-character MAX_PATH limitation
on Windows. However this limitation can be configured away with this
command:

  git config --global core.longpaths true

This is useful when cloning a repo that contains long paths, or if your
repo root has a longer path than expected.

Trivia: even with this configured the longest repo root path is around
210 characters. Any longer than that and "fatal: '$GIT_DIR' too big"
will be hit.

This issue was first hit and was initially tested with
"fetch infra_internal" but could hit elsewhere.

Testing has confirmed that this config setting lets git handle repos
and repo base paths that it otherwise fails on. Setting the
LongPathsEnabled registry key, in contrast, does not work.

Bug: 1472848
Change-Id: I20d6c1d944987ce44d6f7d5cb84a53f1162b8cf9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5046121
Reviewed-by: Solomon Kinard <solomonkinard@chromium.org>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1227793}