[Files]: Separating callback from other parameters.
RecentSource uses an approach where parameters given to GetRecentFiles
are stored in the optional field. That field is reset when the recent
source is done with collecting files. If it is done because the scan
finished, this is fine. If it is done because the timeout occurred, it
is possible that whoever holds to those params gets a null pointer,
which causes a crash.

This CL moves the callback argument out of params. This was what caused
the params to be forced to be move-only type. By moving callback out we
can make the params a copyable object. Thus early termination of the
slow will not empty the params for stragglers.

Bug: b:311934766
Change-Id: I357133b6dc2ec45c1673c59b08c17067f7a333db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5051783
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Bo Majewski <majewski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1227815}