[//net] Take out DwC() from HttpAuthSSPI::GetNextSecurityToken
This DumpWithoutCrash() has been going off at a high volume for a long
time, which pollutes crash reports. Not replacing it with a DCHECK() as
that's against the spirit of DCHECKs() - DCHECKs() are supposed to
represent invariants, but we *know* that this code is actually entered
in production.

Bug: 1466405
Change-Id: I68d63f6709a8637d5bcb1796030e521c6453935a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5049900
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1227856}