[Waffle][iOS] Deregister the promo after a default was set
The conditions on whether or not to display the choice screen promo are
only checked on Chrome startup. This ensures that the promo gets
deregistered when it should, regardless of exactly when conditions are
checked.

This CL also fixes the font of the primary button.

Fixed: b/312432706, b/312411172
Change-Id: Ide0d51729cd12738229a6d98c9c2e65edf9619f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5049860
Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
Commit-Queue: Samar Chehade-Lepleux <samarchehade@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1227866}