Unload: Disable DeprecateUnload for SuddenTerminationDisabler tests
This tests are checking if the `Navigation.SuddenTerminationDisabler`
UMA metrics are correctly recorded for pages with unload handlers.
Since the metrics were added to understand the unload handler usage,
we only need this test when the unload is not deprecated.

This CL disables DeprecateUnload for the tests.

Bug: 1488371
Change-Id: Ia3fbf197d2590585eaaf08b1fd176ba0309fa259
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5039878
Reviewed-by: Fergal Daly <fergal@chromium.org>
Commit-Queue: Mingyu Lei <leimy@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1227948}