Add infrastructure to maintain the type used to fill an AutofilledField
Currently, Autofill only uses the classified type of a field in order to
fill it. In the future, this might change and we might sometimes
fallback to other types in case the classified type doesn't lead to a
successful filling.

In that case, `AutofillField::overall_type_` stops representing
accurately the filling state of the field, which might be needed by
other components.

This CL:
- Adds an `autofilled_type_` optional data member to AutofillField
tracking the type used to fill that field by Autofill.
- Adds logic to retrieve this information from the cache since this
information does not come from the renderer and hence newly built
AutofillFields cannot assign it.
- Adds logic to maintain history of that data member so that Undo
operations keeps this information consistent with the other states of
the field.
- Adds minimal logic to set this type whenever a field is autofilled.

Logic that will compute fallback types is out of scope of this CL.

This CL is a no-op.

Bug: b/311604770
Change-Id: Ie88f7fea391ea716ee1608822490e327c9b1354f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5048953
Reviewed-by: Matthias Körber <koerber@google.com>
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1227965}