[ios] Improve use of __weak SceneState* in TestBrowser
When SceneState is created in the constructor it is immediately
destroyed (because the property is weak), for this reason scene_state_
will be nil at the end of the constructor. So there is no need to
create the SceneState in those cases.

Note that this change created no regression in tests, so no additional
fix is needed.

Bug: none
Change-Id: Iedd9af0e2a5773e969778a8c04eeddf364850b85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5049871
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Auto-Submit: Federica Germinario <fedegermi@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1227982}