picture-in-picture: Add PictureInPictureOcclusionTracker
This CL adds a PictureInPictureOcclusionTracker, which keeps track of
picture-in-picture widgets and widgets that need to know when they're
occluded by a picture-in-picture window, and notifies observers when
the occlusion state changes.

This will be used to ensure that security-sensitive UI does not get
covered up by a picture-in-picture window, which would leave the user
vulnerable to spoofing.

This CL just adds the tracker, but does not register any
picture-in-picture widgets to it nor any observers.

Bug: 1498673
Change-Id: I24b554ded8eeb39a76f6659d99076184f4fd539a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5041595
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1227983}