[PwaRestore]: Polish the UI a bit.
This is mostly a subtle changes in margins and such, but also the
background for the apps is a little bit darker (it was a bit hard
to notice the background) and the placeholder icons now have a
letter (instead of a question mark).

This brings the implementation a bit closer to the UX spec.
Some work remains though.

Bug: 1481570
Change-Id: I404a21b68eb6aad571f7ce9ed25d593b19cf4cae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5050703
Reviewed-by: Glenn Hartmann <hartmanng@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Finnur Thorarinsson <finnur@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1227987}