[IdleTimeout] Adding actions for IdleTimeoutActions on iOS - Part1
This CL implements the ActionRunenr for the IdleTimeoutActions policy. The actions run in sequence based on priority, and if one actions
fails, the following actions do not run.

The actions that can run are clearing data, signing out and closing
tabs. Only the action for clearing data is implemented in this CL.

Note: The IdleTimeoutActions.yaml file will only updated with
close_tabs and sign_out once the actions and UI are fully implemented and tested.

Bug: b/301676922
Change-Id: I22f73b558bb673f6f74677fc025e5d7fb41945a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5024656
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Commit-Queue: Salma Elmahallawy <esalma@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1227988}