Consistently clear sync metadata on SyncMode changes
Sync metadata should usually be cleared when Sync is stopped, and any
left-behind metadata should eventually get cleaned up - and after
crbug.com/897628, that was generally the case. But there was one edge
case that wasn't handled yet: When the Sync machinery missed a switch
between "full Sync" (aka Sync-the-feature) and transport mode, the
metadata of the "other" (now-inactive) mode would be left behind and
not cleaned up.
This CL fixes this by doing two things:
* Adding ClearMetadataWhileStopped() calls (on both delegates i.e.
  processors) in ModelLoadManager::LoadModels(), after the correct
  delegate was started.
* Changing the semantics of ClearMetadataWhileStopped() so that it's
  a no-op if the type in question is not actually stopped.
Together, this means the metadata of the "wrong" mode gets cleared,
without affecting the metadata of the "correct" mode.

Bug: 1502832
Change-Id: I41de9c4d9720efa2caf0dc94493775b777265460
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5033412
Reviewed-by: Ankush Singh <ankushkush@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1227997}