Remove partial support for proxy chains in PAC strings
For the moment, the plan is to not support proxy chains in PAC strings.
This CL rewrites a bunch of uses of `{ProxyInfo,ProxyList}.ToPacString`
to use `ToDebugString` (which *can* represent a chain), and removes
TODO's for supporting parsing PAC strings into chains.

Three spots that call ToPacString to return the result of proxy
resolution externally are updated to fail if that result contains a
multi-proxy chain.

Bug: 1491092
Change-Id: I30cdef9d3293312b0c52218f0e0baba639bcb76d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5034180
Reviewed-by: Adam Rice <ricea@chromium.org>
Reviewed-by: Roland Bock <rbock@google.com>
Commit-Queue: Dustin Mitchell <djmitche@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Derek Schuff <dschuff@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1228039}