[Site search policy] Support for site search in TemplateURL
This CL makes the following changes to `TemplateURL` for the new
SiteSearchSettings Enterprise policy:
- Add a new field `featured_by_policy` to indicate when a SE should
  be featured in the Omnibox
- Changes the type of field `created_by_policy` from a boolean to
  an enum to represent which policy introduced the SE:
  - 0 = no policy
  - 1 = Default Search Provider policy
  - 2 = Site Search policy

This is the downstream effect in the keywords DB table:
- Add a new column corresponding to `featured_by_policy`
- Change the code that reads and writes entries to the table to
  convert to/from the enum instead of a boolean

There is no need to change the table schema for the `created_by_policy`
field, because the corresponding column is already an integer with
the same representation as above (0: no policy, 1: DSP policy).

Screenshots: b/307544344#comment5

We were originally planning to only add `featured_by_policy` to `TemplateURL`, but unfortunately that doesn't allow us to
differentiate from the default search provider set by policy:

- Today, there are two booleans to represent policy state:
  `created_by_policy` (DSP policy) and `enforced_by_policy` (whether
  the policy is required or recommended)
- Those two booleans represent 3 valid states: no policy, policy is
  mandatory, policy is recommended
- The site search policy is mandatory only in the initial version, but
  we may eventually support the recommended state -> that will require
  us to support more states than currently possible

Please refer to crrev.com/c/5017736 for a discussion on alternatives.

Bug: b/309457587, b/307544344
Change-Id: I47c8411ead3a78dbc9519114b50a6f04777b6188
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5017083
Commit-Queue: Fabio Tirelo <ftirelo@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1228047}