Let MenuModelAdapter::CreateMenu return a std::unique_ptr<>.
This allows to eliminate the remaining instances of the MenuRunner
constructor that accepts a MenuItemView*.

As a side-task, this CL cleans up some dangling pointers to the
MenuItemViews created via MenuModelAdapter::CreateMenu. They are
almost all due to incorrect definition orders. Since the runner
(indirectly) owns the `MenuItemView`, it must outlive it. Therefore
the member variable corresponding to the runner should be listed
before the view.

Bug: 1501294
Change-Id: I5612db08a22eddbf3d23c84afd30c348d24f7d64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5048950
Commit-Queue: Jan Keitel <jkeitel@google.com>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1228065}