[PWA/IPH] Set up IPH feature services for link capturing bubble
This CL sets up the IPH services to show a custom action IPH bubble
that redirects to the app settings page if accepted.

Currently, the IPH is configured to always show up, and will be
controlled by guardrails implemented in the web_app system.

Bug: 1412856, b/302136219
Change-Id: I6f9bd60ddfe889ce7b80762c6af21b8c50fbc883
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5031337
Reviewed-by: Dana Fried <dfried@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1228070}