[UPMLocalSettings] LaunchCheckupOnDevice takes the account as argument
The syncing account is passed from ChromePasswordManagerClient to
CredentialLeakControllerAndroid and PasswordCheckupLauncherHelperImpl
such that it could be used in Java to decide to show the Password
Check UI for the account or local. Passing to Java and showing the UI
comes in a follow-up.

I've passed a new parameter to LaunchCheckupOnDevice (there is change
in logic).

Bug: b/306669939, 1504180
Change-Id: Ia2c901ff53f5592718d54e6dcaf2afea4503dda8
Coverage-Reason: Pre-existing lack of test coverage in the lines where
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5034924
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Commit-Queue: Ivana Žužić <izuzic@google.com>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1228098}