Disable deprecated non-streaming declarative shadow DOM
This patch disables the feature. It was deprecated via the I2S below:

https://groups.google.com/a/chromium.org/g/blink-dev/c/1n0MLr8G5NQ/m/ZcVvTd_vAgAJ

This feature has been slowly being disabled on stable Chrome starting
with M119. As of 11-22-23 it is disabled for 2% of stable, and no
bugs have been reported. This CL disables it in code starting in
M121, but that will be reverted if problems are encountered.

The use counter is here:

https://chromestatus.com/metrics/feature/timeline/popularity/3196

It is likely that most of this usage is "safe" to deprecate, because
the content has fallbacks for the case that the feature is not
supported. This feature was never supported in Gecko or WebKit, so
fallback code is highly likely.

Bug: 1396384
Change-Id: I5c53d0ceba013af6769e017c257026f77edff4b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4776313
Auto-Submit: Mason Freed <masonf@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1228104}