De-flake input-text-double-click.html
This test is flaking on mac because sometimes the input element is not
focused when the screenshot is taken.

This patch changes the test to just look at the selection api in script
instead of using screenshots to determine that double clicking the input
selects the correct text.

Fixed: 1501093
Change-Id: Icabf15b94158914b124f3c970516fa59808833b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5021057
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Reviewed-by: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1228146}