evaluate_seed: handle fieldtrial_testing_config.
Previously, we'd bail out early if SetUpFieldTrials returned false, but
it will do that if the fieldtrial_testing_config.json is applied. So,
do *not* bail early in this case, and instead continue execution to
serialize whatever state is present in the global FeatureList. (e.g.
the evaluated fieldtrial_testing_config state).

BUG=b:311739278
TEST=/opt/google/chrome/evaluate_seed --enable-field-trial-config

Change-Id: I3b0432a403113e53cb1b652fe22e97b9f599969c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5047054
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1228212}