Fix renderer content settings (part 8 / N)
This CL adds a test to verify that extensions can use the content
settings API to block images based on the origin of the resource. This
is functionality that is otherwise not available in the product, and
there was no previous test coverage.

Change-Id: Ia42b3b57712c36a57fcba9a7073c89c4a14599ba
Bug: 1208547
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5047587
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1228220}